Added a demonstration mode

+++ routes/blacklist.py
@@ -1,9 +1,8 @@
-from sqlite3.dbapi2 import SQLITE_ALTER_TABLE
 import flask
+import time
 
 import globals
 
-from utilities.persistence import upsert_persistence
 from classes.sqlite import SQLite
 
 def blacklist_page():
@@ -16,14 +15,18 @@ def blacklist_page():
 	they_sent = []
 	we_sent = []
 
-	# separate the ones that sent and the ones we sent
-	for blacklist in db.get_blacklist():
-		insert = {'shaketag': blacklist[0], 'amount': blacklist[1]}
+	if (globals.bot_flags['demo'] == False):
+		# separate the ones that sent and the ones we sent
+		for blacklist in db.get_blacklist():
+			insert = {'shaketag': blacklist[0], 'amount': blacklist[1]}
 
-		if (blacklist[1] > 0):
-			they_sent.append(insert)
-		else:
-			we_sent.append(insert)
+			if (blacklist[1] > 0):
+				they_sent.append(insert)
+			else:
+				we_sent.append(insert)
+	else:
+		they_sent = [{'shaketag': '@unclejimmy', 'amount': '25.49'}]
+		we_sent = [{'shaketag': '@gasmoney', 'amount': '-10.00'}]
 
 	data = {
 		'version': globals.version,
@@ -55,6 +58,7 @@ def blacklist_add(shaketag):
 		db.upsert_blacklist(shaketag.lower(), amount)
 		db.commit()
 
+	if (data) or (globals.bot_flags['demo'] == True):
 		status_code = 201
 
 	db.close()
@@ -66,6 +70,10 @@ def blacklist_delete(shaketag):
 	Route to remove shaketag from blacklist, will not error out if shaketag does not exist
 	'''
 	
+	if (globals.bot_flags['demo'] == True):
+		time.sleep(1.5)
+		return flask.Response(status = 201)
+
 	db = SQLite()
 	db.delete_blacklist(shaketag.lower())
 	db.commit()
